move key_provider to dedicated package
